sem1/hw1/requirements.txt
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
requests>=2.31.0
accelerate>=0.20.0